requirements.txt
click==8.1.3
ExifRead==2.1.2
Pillow==9.4.0
